Clustering by K-Means
In this script I am running a k-means clustering being a method of vector quantization, originally from signal processing, that aims to partition n observations into k clusters in which each observation belongs to the cluster with the nearest mean (cluster centroids or cluster centroid), serving as a prototype of the cluster. Along this code you will be able to find the most effective number of clusters based on the wss and bss plus an interesting analysis with the silhouette.